+++ src/controller/middleware/fallback.rs
@@ -1,12 +1,16 @@
 //! Fallback Middleware
 //!
-//! This middleware handles fallback logic for the application when routes do not match. It serves
-//! a file, a custom not-found message, or a default HTML fallback page based on the configuration.
+//! This middleware handles fallback logic for the application when routes do
+//! not match. It serves a file, a custom not-found message, or a default HTML
+//! fallback page based on the configuration.
 
-use crate::{app::AppContext, controller::middleware::MiddlewareLayer, Error, Result};
 use axum::{http::StatusCode, response::Html, Router as AXRouter};
-use serde::{Deserialize, Serialize};
+use serde::{Deserialize, Deserializer, Serialize, Serializer};
 use tower_http::services::ServeFile;
+
+use crate::{app::AppContext, controller::middleware::MiddlewareLayer, Result};
+
+pub struct StatusCodeWrapper(pub StatusCode);
 
 #[derive(Default, Debug, Clone, Deserialize, Serialize)]
 pub struct Fallback {
@@ -15,7 +19,12 @@ pub struct Fallback {
     pub enable: bool,
     /// For the unlikely reason to return something different than `404`, you
     /// can set it here
-    pub code: Option<u16>,
+    #[serde(
+        default = "default_status_code",
+        serialize_with = "serialize_status_code",
+        deserialize_with = "deserialize_status_code"
+    )]
+    pub code: StatusCode,
     /// Returns content from a file pointed to by this field with a `404` status
     /// code.
     pub file: Option<String>,
@@ -24,6 +33,33 @@ pub struct Fallback {
     pub not_found: Option<String>,
 }
 
+fn default_status_code() -> StatusCode {
+    StatusCode::OK
+}
+
+fn deserialize_status_code<'de, D>(de: D) -> Result<StatusCode, D::Error>
+where
+    D: Deserializer<'de>,
+{
+    let code: u16 = Deserialize::deserialize(de)?;
+    StatusCode::from_u16(code).map_or_else(
+        |_| {
+            Err(serde::de::Error::invalid_value(
+                serde::de::Unexpected::Unsigned(u64::from(code)),
+                &"a value between 100 and 600",
+            ))
+        },
+        Ok,
+    )
+}
+
+#[allow(clippy::trivially_copy_pass_by_ref)]
+fn serialize_status_code<S>(status: &StatusCode, ser: S) -> Result<S::Ok, S::Error>
+where
+    S: Serializer,
+{
+    ser.serialize_u16(status.as_u16())
+}
 impl MiddlewareLayer for Fallback {
     /// Returns the name of the middleware
     fn name(&self) -> &'static str {
@@ -35,30 +71,23 @@ impl MiddlewareLayer for Fallback {
         self.enable
     }
 
+    fn config(&self) -> serde_json::Result<serde_json::Value> {
+        serde_json::to_value(self)
+    }
+
     /// Applies the fallback middleware to the application router.
     fn apply(&self, app: AXRouter<AppContext>) -> Result<AXRouter<AppContext>> {
         let app = if let Some(path) = &self.file {
             app.fallback_service(ServeFile::new(path))
         } else if let Some(not_found) = &self.not_found {
             let not_found = not_found.to_string();
-            let code = self
-                .code
-                .map(StatusCode::from_u16)
-                .transpose()
-                .map_err(|e| Error::Message(format!("{e}")))?
-                .unwrap_or(StatusCode::NOT_FOUND);
-            app.fallback(move || async move { (code, not_found) })
+            let status_code = self.code;
+            app.fallback(move || async move { (status_code, not_found) })
         } else {
-            let code = self
-                .code
-                .map(StatusCode::from_u16)
-                .transpose()
-                .map_err(|e| Error::Message(format!("{e}")))?
-                .unwrap_or(StatusCode::NOT_FOUND);
             let content = include_str!("fallback.html");
-            app.fallback(move || async move { (code, Html(content)) })
+            let status_code = self.code;
+            app.fallback(move || async move { (status_code, Html(content)) })
         };
-        tracing::info!("[Middleware] +fallback");
         Ok(app)
     }
 }