+++ src/controller/middleware/logger.rs
@@ -27,6 +27,7 @@ impl Default for Config {
 }
 
 /// [`Middleware`] struct responsible for logging HTTP requests.
+#[derive(Serialize)]
 pub struct Middleware {
     config: Config,
     environment: Environment,
@@ -50,6 +51,10 @@ impl MiddlewareLayer for Middleware {
     /// Returns whether the middleware is enabled or not
     fn is_enabled(&self) -> bool {
         self.config.enable
+    }
+
+    fn config(&self) -> serde_json::Result<serde_json::Value> {
+        serde_json::to_value(self)
     }
 
     /// Applies the logger middleware to the application router by adding layers for: