+++ src/controller/middleware/mod.rs
@@ -1,9 +1,10 @@
 //! Base Middleware for Loco Application
 //!
 //! This module defines the various middleware components that Loco provides.
-//! Each middleware is responsible for handling different aspects of request processing, such as
-//! authentication, logging, CORS, compression, and error handling. The middleware can be easily
-//! configured and applied to the application's router.
+//! Each middleware is responsible for handling different aspects of request
+//! processing, such as authentication, logging, CORS, compression, and error
+//! handling. The middleware can be easily configured and applied to the
+//! application's router.
 
 #[cfg(all(feature = "auth_jwt", feature = "with-db"))]
 pub mod auth;
@@ -21,9 +22,11 @@ pub mod request_id;
 pub mod secure_headers;
 pub mod static_assets;
 pub mod timeout;
-use crate::{app::AppContext, Result};
+
 use axum::Router as AXRouter;
 use serde::{Deserialize, Serialize};
+
+use crate::{app::AppContext, Result};
 
 /// Trait representing the behavior of middleware components in the application.
 pub trait MiddlewareLayer {
@@ -35,7 +38,14 @@ pub trait MiddlewareLayer {
         true
     }
 
-    /// Applies the middleware to the given Axum router and returns the modified router.
+    /// Returns middleware config.
+    ///
+    /// # Errors
+    /// when could not convert middleware to [`serde_json::Value`]
+    fn config(&self) -> serde_json::Result<serde_json::Value>;
+
+    /// Applies the middleware to the given Axum router and returns the modified
+    /// router.
     ///
     /// # Errors
     ///
@@ -43,10 +53,12 @@ pub trait MiddlewareLayer {
     fn apply(&self, app: AXRouter<AppContext>) -> Result<AXRouter<AppContext>>;
 }
 
-/// Constructs a default stack of middleware for the Axum application based on the provided context.
+/// Constructs a default stack of middleware for the Axum application based on
+/// the provided context.
 ///
-/// This function initializes and returns a vector of middleware components that are commonly used
-/// in the application. Each middleware is created using its respective `new` function and
+/// This function initializes and returns a vector of middleware components that
+/// are commonly used in the application. Each middleware is created using its
+/// respective `new` function and
 #[must_use]
 pub fn default_middleware_stack(ctx: &AppContext) -> Vec<Box<dyn MiddlewareLayer>> {
     vec![
@@ -72,41 +84,52 @@ pub fn default_middleware_stack(ctx: &AppContext) -> Vec<Box<dyn MiddlewareLayer
 /// Server middleware configuration structure.
 #[derive(Default, Debug, Clone, Deserialize, Serialize)]
 pub struct Config {
-    /// Middleware that enable compression for the response.
+    /// Compression for the response.
     #[serde(default)]
     pub compression: compression::Compression,
-    /// Middleware that enable etag cache headers.
+
+    /// Etag cache headers.
     #[serde(default)]
     pub etag: etag::Etag,
-    /// Middleware that limit the payload request.
+
+    /// Limit the payload request.
     #[serde(default)]
     pub limit_payload: limit_payload::LimitPayload,
-    /// Middleware that improve the tracing logger and adding trace id for each
-    /// request.
+
+    /// Logger and augmenting trace id with request data
     #[serde(default)]
     pub logger: logger::Config,
-    /// catch any code panic and log the error.
+
+    /// Catch any code panic and log the error.
     #[serde(default)]
     pub catch_panic: catch_panic::CatchPanic,
-    /// Setting a global timeout for the requests
+
+    /// Setting a global timeout for requests
     #[serde(default)]
     pub timeout_request: timeout::TimeOut,
-    /// Setting cors configuration
+
+    /// CORS configuration
     #[serde(default)]
     pub cors: cors::Cors,
+
     /// Serving static assets
     #[serde(rename = "static")]
     #[serde(default)]
     pub static_assets: static_assets::StaticAssets,
+
     /// Sets a set of secure headers
     #[serde(default)]
     pub secure_headers: secure_headers::SecureHeader,
+
     /// Calculates a remote IP based on `X-Forwarded-For` when behind a proxy
     #[serde(default)]
     pub remote_ip: remote_ip::RemoteIpMiddleware,
+
     /// Configure fallback behavior when hitting a missing URL
     #[serde(default)]
     pub fallback: fallback::Fallback,
+
+    /// Request ID
     #[serde(default)]
     pub request_id: request_id::RequestId,
 }