+++ src/controller/middleware/static_assets.rs
@@ -1,17 +1,21 @@
 //! Static Assets Middleware.
 //!
-//! This middleware serves static files (e.g., images, CSS, JS) from a specified folder to the client.
-//! It also allows configuration of a fallback file to serve in case a requested file is not found.
-//! Additionally, it can serve precompressed files if enabled via the configuration.
+//! This middleware serves static files (e.g., images, CSS, JS) from a specified
+//! folder to the client. It also allows configuration of a fallback file to
+//! serve in case a requested file is not found. Additionally, it can serve
+//! precompressed files if enabled via the configuration.
 //!
-//! The middleware checks if the specified folder and fallback file exist, and if either is missing,
-//! it returns an error. If the files exist, the middleware is added to the router to serve static files.
+//! The middleware checks if the specified folder and fallback file exist, and
+//! if either is missing, it returns an error. If the files exist, the
+//! middleware is added to the router to serve static files.
 
-use crate::{app::AppContext, controller::middleware::MiddlewareLayer, Error, Result};
+use std::path::PathBuf;
+
 use axum::Router as AXRouter;
 use serde::{Deserialize, Serialize};
-use std::path::PathBuf;
 use tower_http::services::{ServeDir, ServeFile};
+
+use crate::{app::AppContext, controller::middleware::MiddlewareLayer, Error, Result};
 
 /// Static asset middleware configuration
 #[derive(Default, Debug, Clone, Deserialize, Serialize)]
@@ -40,7 +44,7 @@ pub struct FolderConfig {
 impl MiddlewareLayer for StaticAssets {
     /// Returns the name of the middleware.
     fn name(&self) -> &'static str {
-        "static assets"
+        "static_assets"
     }
 
     /// Checks if the static assets middleware is enabled.
@@ -48,15 +52,19 @@ impl MiddlewareLayer for StaticAssets {
         self.enable
     }
 
+    fn config(&self) -> serde_json::Result<serde_json::Value> {
+        serde_json::to_value(self)
+    }
+
     /// Applies the static assets middleware to the application router.
     ///
-    /// This method wraps the provided [`AXRouter`] with a service to serve static files from the
-    /// folder specified in the configuration. It will serve a fallback file if the requested file
-    /// is not found, and can also serve precompressed (gzip) files if enabled.
+    /// This method wraps the provided [`AXRouter`] with a service to serve
+    /// static files from the folder specified in the configuration. It will
+    /// serve a fallback file if the requested file is not found, and can
+    /// also serve precompressed (gzip) files if enabled.
     ///
-    /// Before applying, it checks if the folder and fallback file exist. If either is missing,
-    /// it returns an error.
-    ///
+    /// Before applying, it checks if the folder and fallback file exist. If
+    /// either is missing, it returns an error.
     fn apply(&self, app: AXRouter<AppContext>) -> Result<AXRouter<AppContext>> {
         if self.must_exist
             && (!PathBuf::from(&self.folder.path).exists()