+++ src/controller/middleware/timeout.rs
@@ -34,6 +34,10 @@ impl MiddlewareLayer for TimeOut {
         self.enable
     }
 
+    fn config(&self) -> serde_json::Result<serde_json::Value> {
+        serde_json::to_value(self)
+    }
+
     /// Applies the timeout middleware to the application router.
     ///
     /// This method wraps the provided [`AXRouter`] in a [`TimeoutLayer`], ensuring